git-svn-id: http://plugins.svn.wordpress.org/user-access-manager/trunk@262013 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ class/UserAccessManager.class.php
@@ -413,8 +413,14 @@ class UserAccessManager
         $wud = wp_upload_dir();
         if (empty($wud['error'])) {
             $url = $wud['basedir'] . "/";
-            unlink($url . ".htaccess");
-            unlink($url . ".htpasswd");
+            
+            if (file_exists($url.".htaccess")) {
+                unlink($url.".htaccess");
+            }
+            
+            if (file_exists($url.".htpasswd")) {
+                unlink($url.".htpasswd");
+            }
         }
     }
